Moved auth types to auth.dto
