+++ src/core/dto/auth.dto.ts
@@ -1,9 +1,32 @@
 import "reflect-metadata";
 
-import { IsEmail, IsInt, IsJWT, Length } from "class-validator";
+import {
+  IsEmail,
+  IsInt,
+  IsJWT,
+  IsNotEmpty,
+  IsStrongPassword,
+  Length,
+} from "class-validator";
 import { Field, InputType, ObjectType } from "type-graphql";
 
 import { User } from "../entities/user.entity.js";
+
+interface StrongPasswordOptions {
+  minLength?: number;
+  minLowercase?: number;
+  minUppercase?: number;
+  minNumbers?: number;
+  minSymbols?: number;
+}
+
+export const defaultPasswordOpts: StrongPasswordOptions = {
+  minLength: 8,
+  minLowercase: 1,
+  minUppercase: 1,
+  minNumbers: 1,
+  minSymbols: 1,
+};
 
 @ObjectType()
 export class TokensPayload {
@@ -58,4 +81,33 @@ export class AccountResponse {
 
   @Field()
   message!: string;
+}
+
+@InputType()
+export class ResetPasswordInput {
+  @Field()
+  @IsEmail()
+  email!: string;
+
+  @Field()
+  @Length(8, 53)
+  @IsStrongPassword(defaultPasswordOpts)
+  newPassword!: string;
+}
+
+@InputType()
+export class ChangePasswordInput {
+  @Field()
+  @IsEmail()
+  email!: string;
+
+  @Field()
+  @Length(8, 53)
+  @IsNotEmpty()
+  oldPassword!: string;
+
+  @Field()
+  @Length(8, 53)
+  @IsStrongPassword(defaultPasswordOpts)
+  newPassword!: string;
 }
