+++ src/core/dto/user.dto.ts
@@ -4,28 +4,12 @@ import {
   IsEmail,
   IsNotEmpty,
   IsStrongPassword,
-  Length,
   MaxLength,
 } from "class-validator";
 import { Field, InputType } from "type-graphql";
 
 import { StringScalar } from "../../utils/scalars/string-sanitizer.util.js";
-
-interface StrongPasswordOptions {
-  minLength?: number;
-  minLowercase?: number;
-  minUppercase?: number;
-  minNumbers?: number;
-  minSymbols?: number;
-}
-
-const defaultPasswordOpts: StrongPasswordOptions = {
-  minLength: 8,
-  minLowercase: 1,
-  minUppercase: 1,
-  minNumbers: 1,
-  minSymbols: 1,
-};
+import { defaultPasswordOpts } from "./auth.dto.js";
 
 @InputType()
 export class CreateUserInput {
@@ -42,37 +26,7 @@ export class CreateUserInput {
   @MaxLength(100)
   fullName!: string;
 
-  @Field()
-  @IsNotEmpty()
+  @Field({ nullable: true })
   @MaxLength(15)
-  pronouns!: string;
-}
-
-@InputType()
-export class ResetPasswordInput {
-  @Field()
-  @IsEmail()
-  email!: string;
-
-  @Field()
-  @Length(8, 53)
-  @IsStrongPassword(defaultPasswordOpts)
-  newPassword!: string;
-}
-
-@InputType()
-export class ChangePasswordInput {
-  @Field()
-  @IsEmail()
-  email!: string;
-
-  @Field()
-  @Length(8, 53)
-  @IsNotEmpty()
-  oldPassword!: string;
-
-  @Field()
-  @Length(8, 53)
-  @IsStrongPassword(defaultPasswordOpts)
-  newPassword!: string;
+  pronouns?: string;
 }
